+++ src/main/kotlin/lotto/ui/LottoPurchaseView.kt
@@ -1,9 +1,12 @@
 package lotto.ui
 
-import lotto.LottoSeller
 import lotto.domain.Lotto
+import lotto.domain.LottoSeller
 
 object LottoPurchaseView {
+
+    private val lottoSeller = LottoSeller()
+
     fun inputPriceForPurchase(): Long {
         println("구입금액을 입력해 주세요.")
         val money = readLine()!!.toLong()
@@ -13,7 +16,7 @@ object LottoPurchaseView {
     }
 
     fun purchase(money: Long): List<Lotto> {
-        val lottoList = LottoSeller.purchaseAuto(money)
+        val lottoList = lottoSeller.purchaseAuto(money)
         println("${lottoList.size}개를 구매했습니다.")
         return lottoList
     }
